backend/requirements.txt
flask
flask-cors
python-dotenv
flask-mail
pillow
requests
deta